requirements.txt
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
filelock==3.9.0
Flask==2.2.2
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
Pillow==10.0.1
pip==23.3
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
PyYAML==6.0.1
requests==2.31.0
setuptools==68.0.0
sympy==1.11.1
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.7.1
urllib3==1.26.18
Werkzeug==2.2.3
wheel==0.41.2
win-inet-pton==1.1.0
